--- app/controladores/inicio.php
@@ -1,22 +1,0 @@
-<?php
-
-namespace App\controladores;
-
-use App\interfaces\Database;
-
-
-
-class inicio
-{
-    public bool $breadcrumb = false;
-        
-    public function __construct(Database $coneccion)
-    {
-        
-    }
-
-    public function index()
-    {
-        
-    }
-}
