+++ app/controllers/inicio.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\controllers;
+
+use App\interfaces\Database;
+
+
+
+class inicio
+{
+    public bool $breadcrumb = false;
+        
+    public function __construct(Database $coneccion)
+    {
+        
+    }
+
+    public function index()
+    {
+        
+    }
+}
